requirements.txt
tensorflow
numpy
pillow
fastapi
uvicorn
scikit-learn
matplotlib
seaborn
openai
